scobc-v1: add `OVERRIDES` for multiconfig build
Due to behavior changes in lopper, the PMC device-tree no longer
contains IOMMU nodes.
To handle this difference, additional OVERRIDES were introduced so the
device-tree processing can diverge based on the build context.

This is a workaround; ideally, the devicetree-overlay in
device-tree.bbappend should be upstreamed into the SDT itself.

Signed-off-by: KantaTamura <kanta@spacecubics.com>